Add other flags types to params.rs
